Finally got the new Preference* widgets to work.
The idea is to populate one MultiSelect with the current drivers and cars,
and then maintain a second one where I dynamically put all drivers
you have once selected. But it was surprisingly difficult to
figure out basic stuff in Andriod. For examle, the just creating a class and
hoding a reference to it wouldn't work, because theembedded OS/Framework
manages each thread, and the instance I think I'm about to use,
may not exist anymore, or have been put to sleep, etc.